+++ upkie/cpp/spine/AgentInterface.h
@@ -33,7 +33,7 @@ class AgentInterface {
  public:
   /*! Open interface to the agent at a given shared-memory file.
    *
-   * \param[in] name Name of the shared memory file (e.g. "/vulp").
+   * \param[in] name Name of the shared memory file (e.g. "/upkie").
    * \param[in] size Size in bytes.
    *
    * Shared memory objects are available as files in ``/dev/shm``.