+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- **Breaking:** rename the default shared-memory file to ``/upkie``
 - PID balancer: default to hostname for the agent configuration
 
 ### Fixed